Remodule structure, rename struct
