+++ apps/osrm-requester/src/lib.rs
@@ -16,82 +16,11 @@ pub struct Config {
 /// Координата
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
-pub struct Coordinates {
+pub struct Coordinate {
     /// Широта
     latitude: f32,
     /// Долгота
     longitude: f32,
-}
-
-/// Сообщение (интерфейс обмена данными между сервисами)
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct Message<T> {
-    /// Произвольные данные
-    pub data: T,
-    /// Описание
-    pub description: String,
-    /// Время отправки
-    pub send_time: String
-}
-
-/// Данные из OSRM для отправки в очередь
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct OsrmMessageData<'a> {
-    /// Координаты
-    pub coordinates: Vec<[f64; 2]>,
-    /// Маршруты между двумя ключевыми точками
-    pub legs: &'a Vec<OsmrLeg>
-}
-
-/// Данные OSRM
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OsrmData {
-    /// Маршруты
-    pub routes: Vec<OsrmRoute>
-}
-
-/// Маршрут OSRM
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OsrmRoute {
-    /// Геометрия (полигон)
-    pub geometry: String,
-    /// Маршруты между двумя ключевыми точками
-    pub legs: Vec<OsmrLeg>
-}
-
-/// Маршрут между двумя ключевыми точками
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OsmrLeg {
-    /// Дистанция
-    distance: f32,
-    /// Резюмирующий текст
-    summary: String,
-    /// Шаги
-    steps: Vec<OsrmStep>
-}
-
-/// Шаг OSRM
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OsrmStep {
-    /// Маневр
-    maneuver: OsrmManeuver
-}
-
-/// Маневр OSRM
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OsrmManeuver {
-    /// Тип
-    r#type: String,
-    /// Модификатор
-    modifier: Option<String>,
-    /// локация (координаты)
-    location: Vec<f32>,
-    /// Угол до
-    bearing_after: i32,
-    /// Угол после
-    bearing_before: i32
 }
 
 /// Получение конфига
@@ -103,7 +32,7 @@ pub fn get_config() -> Result<Config> {
 }
 
 /// Получение строки координат (склеивание координат через запятую)
-pub fn get_coordinates_str(coordinates_array: Vec<Coordinates>) -> String {
+pub fn get_coordinates_str(coordinates_array: Vec<Coordinate>) -> String {
     coordinates_array
         .into_iter()
         .map(|c| format!("{},{}", c.longitude, c.latitude))
